Refactor Cognito User management classes so that each service class in in its own file. Not only does this follow the style guide, it sets the stage to avoid circular dependency problems.
