+++ src/app/app.module.ts
@@ -3,7 +3,10 @@ import {NgModule} from "@angular/core";
 import {FormsModule} from "@angular/forms";
 import {HttpModule} from "@angular/http";
 import {AppComponent} from "./app.component";
-import {UserRegistrationService, UserLoginService, UserParametersService, CognitoUtil} from "./service/cognito.service";
+import {UserRegistrationService} from "./service/user-registration.service";
+import {UserParametersService} from "./service/user-parameters.service";
+import {UserLoginService} from "./service/user-login.service";
+import {CognitoUtil} from "./service/cognito.service";
 import {routing} from "./app.routes";
 import {HomeComponent, AboutComponent, HomeLandingComponent} from "./public/home.component";
 import {AwsUtil} from "./service/aws.service";
