+++ src/app/service/user-login.service.ts
@@ -1,0 +1,125 @@
+import {Injectable, Inject} from "@angular/core";
+import {DynamoDBService} from "./ddb.service";
+import {CognitoUtil, Callback, CognitoCallback, LoggedInCallback} from "./cognito.service";
+import {CognitoUser, AuthenticationDetails} from "amazon-cognito-identity-js";
+import * as AWS from "aws-sdk/global";
+
+@Injectable()
+export class UserLoginService {
+
+    constructor(public ddb: DynamoDBService, public cognitoUtil: CognitoUtil) {
+    }
+
+    authenticate(username: string, password: string, callback: CognitoCallback) {
+        console.log("UserLoginService: starting the authentication")
+
+        let authenticationData = {
+            Username: username,
+            Password: password,
+        };
+        let authenticationDetails = new AuthenticationDetails(authenticationData);
+
+        let userData = {
+            Username: username,
+            Pool: this.cognitoUtil.getUserPool()
+        };
+
+        console.log("UserLoginService: Params set...Authenticating the user");
+        let cognitoUser = new CognitoUser(userData);
+        console.log("UserLoginService: config is " + AWS.config);
+        cognitoUser.authenticateUser(authenticationDetails, {
+            newPasswordRequired: function(userAttributes, requiredAttributes) {
+              callback.cognitoCallback(`User needs to set password.`, null);
+            },
+            onSuccess: function (result) {
+
+                var logins = {}
+                logins['cognito-idp.' + CognitoUtil._REGION + '.amazonaws.com/' + CognitoUtil._USER_POOL_ID] = result.getIdToken().getJwtToken();
+
+                // Add the User's Id Token to the Cognito credentials login map.
+                AWS.config.credentials = new AWS.CognitoIdentityCredentials({
+                    IdentityPoolId: CognitoUtil._IDENTITY_POOL_ID,
+                    Logins: logins
+                });
+
+                console.log("UserLoginService: set the AWS credentials - " + JSON.stringify(AWS.config.credentials));
+
+                callback.cognitoCallback(null, result);
+
+            },
+            onFailure: function (err) {
+                callback.cognitoCallback(err.message, null);
+            },
+        });
+    }
+
+    forgotPassword(username: string, callback: CognitoCallback) {
+        let userData = {
+            Username: username,
+            Pool: this.cognitoUtil.getUserPool()
+        };
+
+        let cognitoUser = new CognitoUser(userData);
+
+        cognitoUser.forgotPassword({
+            onSuccess: function () {
+
+            },
+            onFailure: function (err) {
+                callback.cognitoCallback(err.message, null);
+            },
+            inputVerificationCode() {
+                callback.cognitoCallback(null, null);
+            }
+        });
+    }
+
+    confirmNewPassword(email: string, verificationCode: string, password: string, callback: CognitoCallback) {
+        let userData = {
+            Username: email,
+            Pool: this.cognitoUtil.getUserPool()
+        };
+
+        let cognitoUser = new CognitoUser(userData);
+
+        cognitoUser.confirmPassword(verificationCode, password, {
+            onSuccess: function () {
+                callback.cognitoCallback(null, null);
+            },
+            onFailure: function (err) {
+                callback.cognitoCallback(err.message, null);
+            }
+        });
+    }
+
+    logout() {
+        console.log("UserLoginService: Logging out");
+        this.ddb.writeLogEntry("logout");
+        this.cognitoUtil.getCurrentUser().signOut();
+
+    }
+
+    isAuthenticated(callback: LoggedInCallback) {
+        if (callback == null)
+            throw("UserLoginService: Callback in isAuthenticated() cannot be null");
+
+        let cognitoUser = this.cognitoUtil.getCurrentUser();
+
+        if (cognitoUser != null) {
+            cognitoUser.getSession(function (err, session) {
+                if (err) {
+                    console.log("UserLoginService: Couldn't get the session: " + err, err.stack);
+                    callback.isLoggedIn(err, false);
+                }
+                else {
+                    console.log("UserLoginService: Session is " + session.isValid());
+                    callback.isLoggedIn(err, session.isValid());
+                }
+            });
+        } else {
+            console.log("UserLoginService: can't retrieve the current user");
+            callback.isLoggedIn("Can't retrieve the CurrentUser", false);
+        }
+    }
+
+}
